Remove the pointless header file
